webservice/requirements.txt

fastapi==0.100.0
pandas==2.0.3
pydantic==2.0.2
uvicorn
SQLAlchemy
psycopg2-binary
